JPA 重构完成

+++ src/main/java/com/moonykun/demo/controller/UserController.java
@@ -68,7 +68,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public String getUser(@PathVariable("id") Long id, Model model) {
+    public String getUser(@PathVariable("id") Integer id, Model model) {
         User user = userService.getUserById(id);
         model.addAttribute("user",user);
         return "user/userEdit";
